+++ src/main/java/com/devbliss/gpullr/controller/GithubUserFetcher.java
@@ -5,25 +5,22 @@ import com.devbliss.gpullr.service.UserService;
 import com.devbliss.gpullr.service.github.GithubApi;
 import com.devbliss.gpullr.util.Log;
 import java.io.IOException;
-import java.text.DateFormat;
 import java.time.Instant;
 import java.time.LocalTime;
+import java.time.temporal.ChronoUnit;
 import java.util.Date;
 import java.util.List;
 import org.slf4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.scheduling.concurrent.ThreadPoolTaskScheduler;
 import org.springframework.stereotype.Component;
 
 /**
  * Fetches all users belonging to Devbliss from GitHub API.
  */
 @Component
-public class GithubUserFetcher {
+public class GithubUserFetcher extends AbstractFixedScheduleFetcher {
 
   private static final int HOURS_OF_DAY = 24;
-
-  private static final int SECONDS_PER_HOUR = 3600;
 
   @Log
   private Logger logger;
@@ -34,21 +31,11 @@ public class GithubUserFetcher {
   @Autowired
   private UserService userService;
 
-  private ThreadPoolTaskScheduler executor;
-
-  public GithubUserFetcher() {
-    executor = new ThreadPoolTaskScheduler();
-    executor.initialize();
-  }
-
-  public void fetchUsers() {
+  @Override
+  protected void fetch() {
     try {
-      logger.info("Start fetching users from GitHub...");
       List<User> users = githubApi.fetchAllOrgaMembers();
       users.forEach(this::handleUser);
-      logger.info("Finished fetching users from GitHub,");
-
-      executor.schedule(() -> fetchUsers(), calculateNextUserFetch());
     } catch (IOException e) {
       logger.error("Error fetching users from GitHub: " + e.getMessage(), e);
     }
@@ -65,15 +52,11 @@ public class GithubUserFetcher {
    *
    * @return
    */
-  private Date calculateNextUserFetch() {
+  @Override
+  protected Date nextFetch() {
     int diff = HOURS_OF_DAY - LocalTime.now().getHour();
     logger.debug("Still " + diff + " hours until midnight.");
     diff++;
-    Date nextExecution = Date.from(Instant.now().plusSeconds(diff * SECONDS_PER_HOUR));
-    logger.debug("The next fetch of organization members from github will be at: "
-        + DateFormat.getInstance().format(nextExecution));
-
-    return nextExecution;
+    return Date.from(Instant.now().plus(diff, ChronoUnit.HOURS));
   }
-
 }