+++ src/main/java/com/devbliss/gpullr/repository/PersistenceConfig.java
@@ -1,10 +1,15 @@
 package com.devbliss.gpullr.repository;
 
+import javax.annotation.PostConstruct;
 import javax.sql.DataSource;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.config.ConfigurableBeanFactory;
+import org.springframework.beans.factory.config.Scope;
 import org.springframework.boot.autoconfigure.jdbc.DataSourceBuilder;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Profile;
+import org.springframework.context.support.SimpleThreadScope;
 import org.springframework.orm.jpa.JpaVendorAdapter;
 import org.springframework.orm.jpa.LocalContainerEntityManagerFactoryBean;
 import org.springframework.orm.jpa.vendor.Database;
@@ -15,13 +20,24 @@ public class PersistenceConfig {
 
   private static final String DRIVER_CLASS_NAME = "org.h2.Driver";
 
+  @Autowired
+  private ConfigurableBeanFactory beanFactory;
+
+  @PostConstruct
+  public void registerThreadScope() {
+    Scope threadScope = new SimpleThreadScope();
+    beanFactory.registerScope("thread", threadScope);
+  }
+
   @Profile("test")
   @Bean
   public DataSource createTestDataSource() {
     return DataSourceBuilder.create().url("jdbc:h2:mem:gpullrDb_test").driverClassName(DRIVER_CLASS_NAME).build();
   }
 
-  @Profile({"prod", "dev"})
+  @Profile({
+      "prod", "dev"
+  })
   @Bean
   public DataSource createDataSource() {
     return DataSourceBuilder.create().url("jdbc:h2:./gpullrDb").driverClassName(DRIVER_CLASS_NAME).build();